refactor/fix: move reducers to separate files to solve warnings
- Extracted the `authReducer`, `Acion`, `State` and `initState` from `AuthContext.tsx` into its own file `authReducer.ts`.
- Extracted the `gameReducer`, `Acion`, `State` and `initState` from `GameContext.tsx` into its own file `gameReducer.ts`.
